[REFACTOR]: 주석 제거, CarName 이름 수정, getter 위치 변경
